Added redux actions to language types file and moved to proper redux directory
