+++ src/client/app/components/RouteComponent.tsx
@@ -19,7 +19,7 @@ import {LinkOptions} from 'actions/graph';
 import {hasToken} from '../utils/token';
 import {showErrorNotification} from '../utils/notifications';
 import {ChartTypes} from '../types/redux/graph';
-import {LanguageTypes} from '../types/i18n';
+import {LanguageTypes} from '../types/redux/i18n';
 import {verificationApi} from '../utils/api';
 import translate from '../utils/translate';
 import {validateComparePeriod, validateSortingOrder} from '../utils/calculateCompare';
