+++ src/client/app/types/items.ts
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { ChartTypes } from '../types/redux/graph';
-import { LanguageTypes } from '../types/i18n';
+import { LanguageTypes } from './redux/i18n';
 
 /**
  * The type of options displayed in Select components.
