+++ src/client/app/types/redux/i18n.ts
@@ -2,8 +2,19 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+import { ActionType } from './actions';
+
 export enum LanguageTypes {
 	en = 'en',
 	fr = 'fr',
 	es = 'es'
+}
+
+export interface ChangeLanguageToRenderAction {
+	type: ActionType.ChangeChartToRender;
+	languageType: LanguageTypes;
+}
+
+export interface GraphState {
+	languageToRender: LanguageTypes;
 }
